Fixed host colour and host ordering tests; added console output for test status.
'draw: host color' now passes because the necessary html elements are postioned
in the correct locations and order on the page.

'draw: host ordering' now passes because it is ambivalent towards host colour,
and relies on host name now

Console outputs whenever a test fails, and produces a success or failure output
on the termination of all tests.